Update: Fixing conflict btw pages and wagtail
