+++ baseapp_wagtail/base/graphql/mutations.py
@@ -1,0 +1,8 @@
+from grapple.schema import schema as grapple_schema
+
+
+try:
+    WagtailMutation = grapple_schema.Mutation
+except AttributeError:
+    class WagtailMutation:
+        pass
