+++ baseapp_wagtail/base/graphql/queries.py
@@ -1,0 +1,5 @@
+from grapple.schema import schema as grapple_schema
+
+
+class WagtailQuery(grapple_schema.Query):
+    pass
