+++ baseapp_wagtail/base/graphql/subscriptions.py
@@ -1,0 +1,8 @@
+from grapple.schema import schema as grapple_schema
+
+
+try:
+    WagtailSubscription = grapple_schema.Subscription
+except AttributeError:
+    class WagtailSubscription:
+        pass
